requirements.txt
Django>=3.2,<3.3
djangorestframework>=3.12.0,<3.13.0
psycopg2>=2.8.6,<2.9
Pillow>=8.3.0,<8.4.0
flake8>=3.9.0,<3.10.0